Remove formatter concurrency to share more code
