+++ src/CodeStyle/CSharp/Analyzers/Microsoft.CodeAnalysis.CSharp.CodeStyle.csproj
@@ -25,6 +25,9 @@
     <Compile Include="..\..\..\Workspaces\CSharp\Portable\Extensions\SyntaxTriviaExtensions.cs" Link="Formatting\SyntaxTriviaExtensions.cs" />
     <Compile Include="..\..\..\Workspaces\CSharp\Portable\Formatting\DefaultOperationProvider.cs" Link="Formatting\DefaultOperationProvider.cs" />
     <Compile Include="..\..\..\Workspaces\CSharp\Portable\Formatting\Engine\AggregatedFormattingResult.cs" Link="Formatting\Engine\AggregatedFormattingResult.cs" />
+    <Compile Include="..\..\..\Workspaces\CSharp\Portable\Formatting\Engine\CSharpFormatEngine.cs" Link="Formatting\Engine\CSharpFormatEngine.cs" />
+    <Compile Include="..\..\..\Workspaces\CSharp\Portable\Formatting\Engine\CSharpStructuredTriviaFormatEngine.cs" Link="Formatting\Engine\CSharpStructuredTriviaFormatEngine.cs" />
+    <Compile Include="..\..\..\Workspaces\CSharp\Portable\Formatting\Engine\FormattingResult.cs" Link="Formatting\Engine\FormattingResult.cs" />
     <Compile Include="..\..\..\Workspaces\CSharp\Portable\Formatting\Engine\Trivia\CSharpTriviaFormatter.cs" Link="Formatting\Engine\Trivia\CSharpTriviaFormatter.cs" />
     <Compile Include="..\..\..\Workspaces\CSharp\Portable\Formatting\Engine\Trivia\CSharpTriviaFormatter.DocumentationCommentExteriorCommentRewriter.cs" Link="Formatting\Engine\Trivia\CSharpTriviaFormatter.DocumentationCommentExteriorCommentRewriter.cs" />
     <Compile Include="..\..\..\Workspaces\CSharp\Portable\Formatting\Engine\Trivia\TriviaDataFactory.Analyzer.cs" Link="Formatting\Engine\Trivia\TriviaDataFactory.Analyzer.cs" />
