+++ src/Workspaces/CSharp/Portable/Formatting/Engine/CSharpFormatEngine.cs
@@ -1,12 +1,9 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.Collections.Generic;
-using Microsoft.CodeAnalysis.CSharp.Symbols;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Formatting.Rules;
 using Microsoft.CodeAnalysis.Options;
-using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.CSharp.Formatting
 {
@@ -22,8 +19,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
                  optionSet,
                  formattingRules,
                  token1,
-                 token2,
-                 TaskExecutor.Concurrent)
+                 token2)
         {
         }
 
@@ -34,7 +30,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
 
         protected override AbstractFormattingResult CreateFormattingResult(TokenStream tokenStream)
         {
-            return new FormattingResult(this.TreeData, tokenStream, this.SpanToFormat, this.TaskExecutor);
+            return new FormattingResult(this.TreeData, tokenStream, this.SpanToFormat);
         }
     }
 }
