+++ src/Workspaces/CSharp/Portable/Formatting/Engine/CSharpStructuredTriviaFormatEngine.cs
@@ -1,14 +1,8 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.Threading;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Symbols;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Options;
-using Microsoft.CodeAnalysis.Text;
-using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.Formatting
 {
@@ -23,7 +17,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
         {
             var root = trivia.GetStructure();
             var formatter = new CSharpStructuredTriviaFormatEngine(trivia, initialColumn, optionSet, formattingRules, root.GetFirstToken(includeZeroWidth: true), root.GetLastToken(includeZeroWidth: true));
-            return formatter.FormatAsync(cancellationToken).WaitAndGetResult_CanCallOnBackground(cancellationToken);
+            return formatter.Format(cancellationToken);
         }
 
         private CSharpStructuredTriviaFormatEngine(
@@ -37,8 +31,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
                  optionSet,
                  formattingRules,
                  token1,
-                 token2,
-                 TaskExecutor.Synchronous)
+                 token2)
         {
         }
 
@@ -52,7 +45,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
             return new FormattingContext(this, tokenStream, LanguageNames.CSharp);
         }
 
-        protected override NodeOperations CreateNodeOperationTasks(CancellationToken cancellationToken)
+        protected override NodeOperations CreateNodeOperations(CancellationToken cancellationToken)
         {
             // ignore all node operations for structured trivia since it is not possible for this to have any impact currently.
             return NodeOperations.Empty;
@@ -60,7 +53,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
 
         protected override AbstractFormattingResult CreateFormattingResult(TokenStream tokenStream)
         {
-            return new FormattingResult(this.TreeData, tokenStream, this.SpanToFormat, this.TaskExecutor);
+            return new FormattingResult(this.TreeData, tokenStream, this.SpanToFormat);
         }
     }
 }
