+++ src/Workspaces/Core/Portable/Formatting/Context/FormattingContext.cs
@@ -268,8 +268,10 @@ namespace Microsoft.CodeAnalysis.Formatting
             var valuePairs = new(SuppressOperation operation, bool shouldSuppress, bool onSameLine)[operations.Count];
 
             // TODO: think about a way to figure out whether it is already suppressed and skip the expensive check below.
-            _engine.TaskExecutor.For(0, operations.Count, i =>
+            for (var i = 0; i < operations.Count; i++)
             {
+                cancellationToken.ThrowIfCancellationRequested();
+
                 var operation = operations[i];
 
                 // if an operation contains elastic trivia itself and the operation is not marked to ignore the elastic trivia
@@ -278,12 +280,12 @@ namespace Microsoft.CodeAnalysis.Formatting
                 {
                     // don't bother to calculate line alignment between tokens
                     valuePairs[i] = (operation, shouldSuppress: false, onSameLine: false);
-                    return;
+                    continue;
                 }
 
                 var onSameLine = _tokenStream.TwoTokensOriginallyOnSameLine(operation.StartToken, operation.EndToken);
                 valuePairs[i] = (operation, shouldSuppress: true, onSameLine: onSameLine);
-            }, cancellationToken);
+            }
 
             valuePairs.Do(v =>
             {
