+++ src/Workspaces/Core/Portable/Formatting/Engine/AbstractFormatEngine.cs
@@ -4,7 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Formatting.Rules;
 using Microsoft.CodeAnalysis.Internal.Log;
 using Microsoft.CodeAnalysis.Options;
@@ -26,8 +25,6 @@ namespace Microsoft.CodeAnalysis.Formatting
     //             that would create too big graph. key for this approach is how to reduce size of graph.
     internal abstract partial class AbstractFormatEngine
     {
-        private const int ConcurrentThreshold = 30000;
-
         private readonly ChainedFormattingRules _formattingRules;
 
         private readonly SyntaxNode _commonRoot;
@@ -37,7 +34,6 @@ namespace Microsoft.CodeAnalysis.Formatting
 
         protected readonly TextSpan SpanToFormat;
 
-        internal readonly TaskExecutor TaskExecutor;
         internal readonly OptionSet OptionSet;
         internal readonly TreeData TreeData;
 
@@ -46,15 +42,13 @@ namespace Microsoft.CodeAnalysis.Formatting
             OptionSet optionSet,
             IEnumerable<IFormattingRule> formattingRules,
             SyntaxToken token1,
-            SyntaxToken token2,
-            TaskExecutor executor)
+            SyntaxToken token2)
             : this(
                   treeData,
                   optionSet,
                   new ChainedFormattingRules(formattingRules, optionSet),
                   token1,
-                  token2,
-                  executor)
+                  token2)
         {
         }
 
@@ -63,13 +57,11 @@ namespace Microsoft.CodeAnalysis.Formatting
             OptionSet optionSet,
             ChainedFormattingRules formattingRules,
             SyntaxToken token1,
-            SyntaxToken token2,
-            TaskExecutor executor)
+            SyntaxToken token2)
         {
             Contract.ThrowIfNull(optionSet);
             Contract.ThrowIfNull(treeData);
             Contract.ThrowIfNull(formattingRules);
-            Contract.ThrowIfNull(executor);
 
             Contract.ThrowIfTrue(treeData.Root.IsInvalidTokenRange(token1, token2));
 
@@ -91,44 +83,34 @@ namespace Microsoft.CodeAnalysis.Formatting
             {
                 _language = token1.Language;
             }
-
-            // set synchronous task executor if it is enabled (explicitly or as part of debug mode) or if there is not
-            // many things to format
-            var synchronousExecutorAllowed =
-                !optionSet.GetOption(FormattingOptions.AllowConcurrent)
-                || optionSet.GetOption(FormattingOptions.DebugMode, _language);
-            var useSynchronousExecutor = synchronousExecutorAllowed || SpanToFormat.Length < ConcurrentThreshold;
-            TaskExecutor = useSynchronousExecutor ? TaskExecutor.Synchronous : executor;
         }
 
         protected abstract AbstractTriviaDataFactory CreateTriviaFactory();
         protected abstract AbstractFormattingResult CreateFormattingResult(TokenStream tokenStream);
 
-        public async Task<AbstractFormattingResult> FormatAsync(CancellationToken cancellationToken)
+        public AbstractFormattingResult Format(CancellationToken cancellationToken)
         {
             using (Logger.LogBlock(FunctionId.Formatting_Format, FormatSummary, cancellationToken))
             {
                 // setup environment
-                var nodeOperations = CreateNodeOperationTasks(cancellationToken);
+                var nodeOperations = CreateNodeOperations(cancellationToken);
 
                 var tokenStream = new TokenStream(this.TreeData, this.OptionSet, this.SpanToFormat, CreateTriviaFactory());
-                var tokenOperationTask = CreateTokenOperationTask(tokenStream, cancellationToken);
+                var tokenOperation = CreateTokenOperation(tokenStream, cancellationToken);
 
                 // initialize context
                 var context = CreateFormattingContext(tokenStream, cancellationToken);
 
                 // start anchor task that will be used later
-                var anchorContextTask = TaskExecutor.ContinueWith(
-                        nodeOperations.AnchorIndentationOperationsTask,
-                        task => task.Result.Do(context.AddAnchorIndentationOperation),
-                        cancellationToken);
+                cancellationToken.ThrowIfCancellationRequested();
+                var anchorContext = nodeOperations.AnchorIndentationOperations.Do(context.AddAnchorIndentationOperation);
 
                 BuildContext(context, tokenStream, nodeOperations, cancellationToken);
 
                 ApplyBeginningOfTreeTriviaOperation(context, tokenStream, cancellationToken);
 
-                await ApplyTokenOperationsAsync(context, tokenStream, anchorContextTask, nodeOperations,
-                    await tokenOperationTask.ConfigureAwait(false), cancellationToken).ConfigureAwait(false);
+                ApplyTokenOperations(context, tokenStream, nodeOperations,
+                    tokenOperation, cancellationToken);
 
                 ApplyTriviaOperations(context, tokenStream, cancellationToken);
 
@@ -147,70 +129,61 @@ namespace Microsoft.CodeAnalysis.Formatting
             return context;
         }
 
-        protected virtual NodeOperations CreateNodeOperationTasks(CancellationToken cancellationToken)
+        protected virtual NodeOperations CreateNodeOperations(CancellationToken cancellationToken)
         {
+            cancellationToken.ThrowIfCancellationRequested();
+
             // iterating tree is very expensive. do it once and cache it to list
-            var nodeIteratorTask = this.TaskExecutor.StartNew(() =>
+            List<SyntaxNode> nodeIterator;
+            using (Logger.LogBlock(FunctionId.Formatting_IterateNodes, cancellationToken))
             {
-                using (Logger.LogBlock(FunctionId.Formatting_IterateNodes, cancellationToken))
+                const int magicLengthToNodesRatio = 5;
+                var result = new List<SyntaxNode>(Math.Max(this.SpanToFormat.Length / magicLengthToNodesRatio, 4));
+
+                foreach (var node in _commonRoot.DescendantNodesAndSelf(this.SpanToFormat))
                 {
-                    const int magicLengthToNodesRatio = 5;
-                    var result = new List<SyntaxNode>(Math.Max(this.SpanToFormat.Length / magicLengthToNodesRatio, 4));
-
-                    foreach (var node in _commonRoot.DescendantNodesAndSelf(this.SpanToFormat))
-                    {
-                        cancellationToken.ThrowIfCancellationRequested();
-                        result.Add(node);
-                    }
-
-                    return result;
+                    cancellationToken.ThrowIfCancellationRequested();
+                    result.Add(node);
                 }
-            },
-            cancellationToken);
+
+                nodeIterator = result;
+            }
 
             // iterate through each operation using index to not create any unnecessary object
-            var indentBlockOperationTask = this.TaskExecutor.ContinueWith(nodeIteratorTask, task =>
+            cancellationToken.ThrowIfCancellationRequested();
+            List<IndentBlockOperation> indentBlockOperation;
+            using (Logger.LogBlock(FunctionId.Formatting_CollectIndentBlock, cancellationToken))
             {
-                using (Logger.LogBlock(FunctionId.Formatting_CollectIndentBlock, cancellationToken))
-                {
-                    return AddOperations<IndentBlockOperation>(task.Result, (l, n) => _formattingRules.AddIndentBlockOperations(l, n), cancellationToken);
-                }
-            },
-            cancellationToken);
+                indentBlockOperation = AddOperations<IndentBlockOperation>(nodeIterator, (l, n) => _formattingRules.AddIndentBlockOperations(l, n), cancellationToken);
+            }
 
-            var suppressOperationTask = this.TaskExecutor.ContinueWith(nodeIteratorTask, task =>
+            cancellationToken.ThrowIfCancellationRequested();
+            List<SuppressOperation> suppressOperation;
+            using (Logger.LogBlock(FunctionId.Formatting_CollectSuppressOperation, cancellationToken))
             {
-                using (Logger.LogBlock(FunctionId.Formatting_CollectSuppressOperation, cancellationToken))
-                {
-                    return AddOperations<SuppressOperation>(task.Result, (l, n) => _formattingRules.AddSuppressOperations(l, n), cancellationToken);
-                }
-            },
-            cancellationToken);
+                suppressOperation = AddOperations<SuppressOperation>(nodeIterator, (l, n) => _formattingRules.AddSuppressOperations(l, n), cancellationToken);
+            }
 
-            var alignmentOperationTask = this.TaskExecutor.ContinueWith(nodeIteratorTask, task =>
+            cancellationToken.ThrowIfCancellationRequested();
+            List<AlignTokensOperation> alignmentOperation;
+            using (Logger.LogBlock(FunctionId.Formatting_CollectAlignOperation, cancellationToken))
             {
-                using (Logger.LogBlock(FunctionId.Formatting_CollectAlignOperation, cancellationToken))
-                {
-                    var operations = AddOperations<AlignTokensOperation>(task.Result, (l, n) => _formattingRules.AddAlignTokensOperations(l, n), cancellationToken);
+                var operations = AddOperations<AlignTokensOperation>(nodeIterator, (l, n) => _formattingRules.AddAlignTokensOperations(l, n), cancellationToken);
 
-                    // make sure we order align operation from left to right
-                    operations.Sort((o1, o2) => o1.BaseToken.Span.CompareTo(o2.BaseToken.Span));
+                // make sure we order align operation from left to right
+                operations.Sort((o1, o2) => o1.BaseToken.Span.CompareTo(o2.BaseToken.Span));
 
-                    return operations;
-                }
-            },
-            cancellationToken);
+                alignmentOperation = operations;
+            }
 
-            var anchorIndentationOperationsTask = this.TaskExecutor.ContinueWith(nodeIteratorTask, task =>
+            cancellationToken.ThrowIfCancellationRequested();
+            List<AnchorIndentationOperation> anchorIndentationOperations;
+            using (Logger.LogBlock(FunctionId.Formatting_CollectAnchorOperation, cancellationToken))
             {
-                using (Logger.LogBlock(FunctionId.Formatting_CollectAnchorOperation, cancellationToken))
-                {
-                    return AddOperations<AnchorIndentationOperation>(task.Result, (l, n) => _formattingRules.AddAnchorIndentationOperations(l, n), cancellationToken);
-                }
-            },
-            cancellationToken);
+                anchorIndentationOperations = AddOperations<AnchorIndentationOperation>(nodeIterator, (l, n) => _formattingRules.AddAnchorIndentationOperations(l, n), cancellationToken);
+            }
 
-            return new NodeOperations(indentBlockOperationTask, suppressOperationTask, anchorIndentationOperationsTask, alignmentOperationTask);
+            return new NodeOperations(indentBlockOperation, suppressOperation, anchorIndentationOperations, alignmentOperation);
         }
 
         private List<T> AddOperations<T>(List<SyntaxNode> nodes, Action<List<T>, SyntaxNode> addOperations, CancellationToken cancellationToken)
@@ -219,8 +192,7 @@ namespace Microsoft.CodeAnalysis.Formatting
             using (var localList = new ThreadLocal<List<T>>(() => new List<T>(), trackAllValues: false))
             {
                 // find out which executor we want to use.
-                var taskExecutor = nodes.Count > (1000 * Environment.ProcessorCount) ? TaskExecutor.Concurrent : TaskExecutor.Synchronous;
-                taskExecutor.ForEach(nodes, n =>
+                foreach (var n in nodes)
                 {
                     cancellationToken.ThrowIfCancellationRequested();
 
@@ -236,7 +208,7 @@ namespace Microsoft.CodeAnalysis.Formatting
                     }
 
                     list.Clear();
-                }, cancellationToken);
+                }
 
                 var operations = new List<T>(localOperations.Values.Sum(v => v.Count));
                 operations.AddRange(localOperations.Values.SelectMany(v => v));
@@ -245,35 +217,34 @@ namespace Microsoft.CodeAnalysis.Formatting
             }
         }
 
-        private Task<TokenPairWithOperations[]> CreateTokenOperationTask(
+        private TokenPairWithOperations[] CreateTokenOperation(
             TokenStream tokenStream,
             CancellationToken cancellationToken)
         {
-            return this.TaskExecutor.StartNew(() =>
+            cancellationToken.ThrowIfCancellationRequested();
+
+            using (Logger.LogBlock(FunctionId.Formatting_CollectTokenOperation, cancellationToken))
             {
-                using (Logger.LogBlock(FunctionId.Formatting_CollectTokenOperation, cancellationToken))
+                // pre-allocate list once. this is cheaper than re-adjusting list as items are added.
+                var list = new TokenPairWithOperations[tokenStream.TokenCount - 1];
+
+                foreach (var pair in tokenStream.TokenIterator)
                 {
-                    // pre-allocate list once. this is cheaper than re-adjusting list as items are added.
-                    var list = new TokenPairWithOperations[tokenStream.TokenCount - 1];
+                    cancellationToken.ThrowIfCancellationRequested();
 
-                    this.TaskExecutor.ForEach(tokenStream.TokenIterator, pair =>
-                    {
-                        var spaceOperation = _formattingRules.GetAdjustSpacesOperation(pair.Item2, pair.Item3);
-                        var lineOperation = _formattingRules.GetAdjustNewLinesOperation(pair.Item2, pair.Item3);
+                    var spaceOperation = _formattingRules.GetAdjustSpacesOperation(pair.Item2, pair.Item3);
+                    var lineOperation = _formattingRules.GetAdjustNewLinesOperation(pair.Item2, pair.Item3);
 
-                        list[pair.Item1] = new TokenPairWithOperations(tokenStream, pair.Item1, spaceOperation, lineOperation);
-                    }, cancellationToken);
-
-                    return list;
+                    list[pair.Item1] = new TokenPairWithOperations(tokenStream, pair.Item1, spaceOperation, lineOperation);
                 }
-            },
-            cancellationToken);
+
+                return list;
+            }
         }
 
-        private async Task ApplyTokenOperationsAsync(
+        private void ApplyTokenOperations(
             FormattingContext context,
             TokenStream tokenStream,
-            Task anchorContextTask,
             NodeOperations nodeOperations,
             TokenPairWithOperations[] tokenOperations,
             CancellationToken cancellationToken)
@@ -281,12 +252,9 @@ namespace Microsoft.CodeAnalysis.Formatting
             var applier = new OperationApplier(context, tokenStream, _formattingRules);
             ApplySpaceAndWrappingOperations(context, tokenStream, tokenOperations, applier, cancellationToken);
 
-            // wait until anchor task to finish adding its information to context
-            await anchorContextTask.ConfigureAwait(false);
-
             ApplyAnchorOperations(context, tokenStream, tokenOperations, applier, cancellationToken);
 
-            await ApplySpecialOperationsAsync(context, tokenStream, nodeOperations, applier, cancellationToken).ConfigureAwait(false);
+            ApplySpecialOperations(context, tokenStream, nodeOperations, applier, cancellationToken);
         }
 
         private void ApplyBeginningOfTreeTriviaOperation(
@@ -342,7 +310,11 @@ namespace Microsoft.CodeAnalysis.Formatting
                 triviaInfo.Format(context, _formattingRules, regularApplier, cancellationToken, tokenPairIndex);
             }
 
-            this.TaskExecutor.For(0, tokenStream.TokenCount - 1, triviaFormatter, cancellationToken);
+            for (var i = 0; i < tokenStream.TokenCount - 1; i++)
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                triviaFormatter(i);
+            }
         }
 
         private TextSpan GetSpanToFormat()
@@ -353,7 +325,7 @@ namespace Microsoft.CodeAnalysis.Formatting
             return TextSpan.FromBounds(startPosition, endPosition);
         }
 
-        private async Task ApplySpecialOperationsAsync(
+        private void ApplySpecialOperations(
             FormattingContext context, TokenStream tokenStream, NodeOperations nodeOperationsCollector, OperationApplier applier, CancellationToken cancellationToken)
         {
             // apply alignment operation
@@ -364,7 +336,7 @@ namespace Microsoft.CodeAnalysis.Formatting
                 // TODO : figure out a way to run alignment operations in parallel. probably find
                 // unions and run each chunk in separate tasks
                 var previousChangesMap = new Dictionary<SyntaxToken, int>();
-                var alignmentOperations = await nodeOperationsCollector.AlignmentOperationTask.ConfigureAwait(false);
+                var alignmentOperations = nodeOperationsCollector.AlignmentOperation;
 
                 alignmentOperations.Do(operation =>
                 {
@@ -390,20 +362,19 @@ namespace Microsoft.CodeAnalysis.Formatting
         {
             using (Logger.LogBlock(FunctionId.Formatting_ApplyAnchorOperation, cancellationToken))
             {
-                var tokenPairsToApplyAnchorOperations = this.TaskExecutor.Filter(
-                                                            tokenOperations,
-                                                            p => AnchorOperationCandidate(p),
-                                                            p => p.PairIndex, cancellationToken);
-
-                cancellationToken.ThrowIfCancellationRequested();
-
                 // TODO: find out a way to apply anchor operation concurrently if possible
                 var previousChangesMap = new Dictionary<SyntaxToken, int>();
-                tokenPairsToApplyAnchorOperations.Do(pairIndex =>
+                foreach (var p in tokenOperations)
                 {
                     cancellationToken.ThrowIfCancellationRequested();
+                    if (!AnchorOperationCandidate(p))
+                    {
+                        continue;
+                    }
+
+                    var pairIndex = p.PairIndex;
                     applier.ApplyAnchorIndentation(pairIndex, previousChangesMap, cancellationToken);
-                });
+                }
 
                 // go through all relative indent block operation, and see whether it is affected by the anchor operation
                 context.GetAllRelativeIndentBlockOperations().Do(o =>
@@ -453,23 +424,13 @@ namespace Microsoft.CodeAnalysis.Formatting
                 var partitioner = new Partitioner(context, tokenStream, tokenOperations);
 
                 // always create task 1 more than current processor count
-                var partitions = partitioner.GetPartitions(this.TaskExecutor == TaskExecutor.Synchronous ? 1 : Environment.ProcessorCount + 1, cancellationToken);
+                var partitions = partitioner.GetPartitions(partitionCount: 1, cancellationToken);
 
-                cancellationToken.ThrowIfCancellationRequested();
-
-                var tasks = new Task[partitions.Count];
-                for (int i = 0; i < partitions.Count; i++)
+                foreach (var partition in partitions)
                 {
-                    var partition = partitions[i];
-                    tasks[i] = this.TaskExecutor.StartNew(() =>
-                    {
-                        cancellationToken.ThrowIfCancellationRequested();
-                        partition.Do(operationPair => ApplySpaceAndWrappingOperationsBody(context, tokenStream, operationPair, applier, cancellationToken));
-                    },
-                    cancellationToken);
+                    cancellationToken.ThrowIfCancellationRequested();
+                    partition.Do(operationPair => ApplySpaceAndWrappingOperationsBody(context, tokenStream, operationPair, applier, cancellationToken));
                 }
-
-                Task.WaitAll(tasks, cancellationToken);
             }
         }
 
@@ -526,10 +487,9 @@ namespace Microsoft.CodeAnalysis.Formatting
             // add scope operation (run each kind sequentially)
             using (Logger.LogBlock(FunctionId.Formatting_BuildContext, cancellationToken))
             {
-                var indentationScopeTask = this.TaskExecutor.ContinueWith(nodeOperations.IndentBlockOperationTask, task => context.AddIndentBlockOperations(task.Result, cancellationToken), cancellationToken);
-                var suppressWrappingScopeTask = this.TaskExecutor.ContinueWith(nodeOperations.SuppressOperationTask, task => context.AddSuppressOperations(task.Result, cancellationToken), cancellationToken);
-
-                Task.WaitAll(new[] { indentationScopeTask, suppressWrappingScopeTask }, cancellationToken);
+                cancellationToken.ThrowIfCancellationRequested();
+                context.AddIndentBlockOperations(nodeOperations.IndentBlockOperation, cancellationToken);
+                context.AddSuppressOperations(nodeOperations.SuppressOperation, cancellationToken);
             }
         }
 
@@ -538,11 +498,10 @@ namespace Microsoft.CodeAnalysis.Formatting
         /// </summary>
         private string FormatSummary()
         {
-            return string.Format("({0}) ({1} - {2}) {3}",
+            return string.Format("({0}) ({1} - {2})",
                 this.SpanToFormat,
                 _token1.ToString().Replace("\r\n", "\\r\\n"),
-                _token2.ToString().Replace("\r\n", "\\r\\n"),
-                this.TaskExecutor.ToString());
+                _token2.ToString().Replace("\r\n", "\\r\\n"));
         }
     }
 }
