+++ src/Workspaces/VisualBasic/Portable/Formatting/Engine/VisualBasicFormatEngine.vb
@@ -1,15 +1,8 @@
 ﻿' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-Imports System.Collections.Generic
-Imports System.Threading
-Imports Microsoft.CodeAnalysis
 Imports Microsoft.CodeAnalysis.Formatting
 Imports Microsoft.CodeAnalysis.Formatting.Rules
 Imports Microsoft.CodeAnalysis.Options
-Imports Microsoft.CodeAnalysis.Text
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Symbols
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
 
 Namespace Microsoft.CodeAnalysis.VisualBasic.Formatting
     Partial Friend Class VisualBasicFormatEngine
@@ -24,8 +17,7 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Formatting
                        optionSet,
                        formattingRules,
                        token1,
-                       token2,
-                       TaskExecutor.Concurrent)
+                       token2)
         End Sub
 
         Protected Overrides Function CreateTriviaFactory() As AbstractTriviaDataFactory
@@ -35,7 +27,7 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Formatting
         Protected Overrides Function CreateFormattingResult(tokenStream As TokenStream) As AbstractFormattingResult
             Contract.ThrowIfNull(tokenStream)
 
-            Return New FormattingResult(Me.TreeData, tokenStream, Me.SpanToFormat, Me.TaskExecutor)
+            Return New FormattingResult(Me.TreeData, tokenStream, Me.SpanToFormat)
         End Function
     End Class
 End Namespace
