+++ src/Workspaces/VisualBasic/Portable/Formatting/Engine/VisualBasicStructuredTriviaFormatEngine.vb
@@ -1,14 +1,8 @@
 ﻿' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-Imports System.Collections.Generic
 Imports System.Threading
-Imports Microsoft.CodeAnalysis
 Imports Microsoft.CodeAnalysis.Formatting
 Imports Microsoft.CodeAnalysis.Options
-Imports Microsoft.CodeAnalysis.Text
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Symbols
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
 
 Namespace Microsoft.CodeAnalysis.VisualBasic.Formatting
     Partial Friend Class VisualBasicStructuredTriviaFormatEngine
@@ -21,7 +15,7 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Formatting
                                       cancellationToken As CancellationToken) As IFormattingResult
             Dim root = trivia.GetStructure()
             Dim formatter = New VisualBasicStructuredTriviaFormatEngine(trivia, initialColumn, optionSet, formattingRules, root.GetFirstToken(includeZeroWidth:=True), root.GetLastToken(includeZeroWidth:=True))
-            Return formatter.FormatAsync(cancellationToken).WaitAndGetResult_CanCallOnBackground(cancellationToken)
+            Return formatter.Format(cancellationToken)
         End Function
 
         Private Sub New(trivia As SyntaxTrivia,
@@ -31,7 +25,7 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Formatting
                        token1 As SyntaxToken,
                        token2 As SyntaxToken)
             MyBase.New(TreeData.Create(trivia, initialColumn),
-                       optionSet, formattingRules, token1, token2, TaskExecutor.Synchronous)
+                       optionSet, formattingRules, token1, token2)
         End Sub
 
         Protected Overrides Function CreateTriviaFactory() As AbstractTriviaDataFactory
@@ -42,13 +36,13 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Formatting
             Return New FormattingContext(Me, tokenStream, LanguageNames.VisualBasic)
         End Function
 
-        Protected Overrides Function CreateNodeOperationTasks(cancellationToken As CancellationToken) As NodeOperations
+        Protected Overrides Function CreateNodeOperations(cancellationToken As CancellationToken) As NodeOperations
             ' ignore all node operations for structured trivia since it is not possible for this to have any impact currently.
             Return NodeOperations.Empty
         End Function
 
         Protected Overrides Function CreateFormattingResult(tokenStream As TokenStream) As AbstractFormattingResult
-            Return New FormattingResult(Me.TreeData, tokenStream, Me.SpanToFormat, Me.TaskExecutor)
+            Return New FormattingResult(Me.TreeData, tokenStream, Me.SpanToFormat)
         End Function
     End Class
 End Namespace
